Divided page into separate components and Dee Man's Corrections
